bootservice/scheduler/requirements.txt
flask
kafka-python
mysql-connector-python
requests
paramiko
azure-storage-file-share
twilio
secure-smtplib